Fix Cloudflare Workers build: isolate bun:sqlite from worker bundle
Split Bun-specific database code (bun:sqlite imports, schema init,
migrations) from server/db/schema.ts into server/db/bun-db.ts so the
CF Workers entry point never transitively imports bun:sqlite.

Also removed jobsRoutes from worker.ts since the in-memory job queue
is Bun-only; CF Workers uses cron triggers instead.
